Bug Fix for Afterlife
We discovered a bug in ghosts' AI which makes them die of hunger and exhaustion even when a relevant building is just next to them. The game can still be enjoyed, but it gets extremely hard and sad at later stages. The updated version fixes the bug; it also now gives a visual hint about where ghosts are going when they're selected.

Old saved games stored in IndexedDB are sadly not compatible with the version of the game fixing the bug. This is due to the minification: each build mangles property names differently. In order to give players a chance to preserve previously built cities, I'd like to expose the buggy build at https://play.js13kgames.com/afterlife/v1.html. It will continue to use the previously saved state, while index.html will create a new one.