Trim some fat
This drops all but 2 games below the 13KB by recompressing them with ECT.
However, in most cases it would've been sufficient to drop DS_Store etc. from within the archives.
In some cases I bundled and minified them as well (not necessary to drop them below the limit - but for the sake of lowering load).

The remaining 2 are both marked as Unfinished, so should hopefully not raise any eyebrows:
- spaceship-repair-factory
- defense-of-the-13th-city